+++ src/util/html.go
@@ -1,0 +1,12 @@
+package util
+
+import "golang.org/x/net/html"
+
+func FindHTMLAttribute(n *html.Node, key string) (string, bool) {
+	for _, attr := range n.Attr {
+		if attr.Key == key {
+			return attr.Val, true
+		}
+	}
+	return "", false
+}
